Decoupling Cache Logger and Statistics, plus a few minor changes in Logging Levels for some outputs
