--- src/tests/utils/stats.log
@@ -1,11 +1,0 @@
-
-----------------------------------
-          Cache Statistics
-----------------------------------
-  Number of cache reads:  1         
-  Number of cache writes: 1         
-  Number of cache hits:   1         
-  Number of cache misses: 1         
-  Cache hit ratio:        50.00000%
-----------------------------------
-
