+++ src/utils/cache_logger.py
@@ -1,81 +1,21 @@
 """
-Description:CacheLogging Utility which handles simulated events, and logs them based on each operation.
+Description: CacheLogging Utility whih handles simulated events, and logs them based on each operation.
 Author: Reece Wayt
 Sources:
     - https://medium.com/@kuldeepkumawat195/python-print-flush-complete-guide-learn-today-f42f87cbc38c
     - https://realpython.com/primer-on-python-decorators/
-
 """
 
 import sys
-from dataclasses import dataclass
 from functools import wraps
 from typing import Callable, TextIO
 
 from common.constants import BusOp, CacheMessage, LogLevel, SnoopResult
 
 
-@dataclass
-class Statistics:
+class CacheLogger:
     """
-    Simple data class used to store essential statistics for simulation. It is directly
-    inherited by the CacheLogger as the Logger is responsible for maintaining, and reporting
-    key statistics of the cache usage.
-
-    Attributes:
-        cache_reads (int): Total number of read attempts to the cache
-        cache_writes (int): Total number of write attempts to the cache
-        cache_hits (int): Number of times requested data was found in cache
-        cache_misses (int): Number of times requested data wasn't found in cache
-
-    Note:
-     As a dataclass, this class automatically generates:
-        - __init__: The constructor method
-        - __repr__: Defines string represention of object
-        - __eq__: Enables equality comparison between instances
-
-    Example:
-        > stats = Statistics()
-        > stats.cache_reads = 100
-        > stats.cache_hits = 75
-        > print(stats.cache_hit_ratio)
-        0.75
-    """
-
-    cache_reads: int = 0
-    cache_writes: int = 0
-    cache_hits: int = 0
-    cache_misses: int = 0
-
-    @property
-    def cache_hit_ratio(self) -> float:
-        """Dynamically calculates hit ratio"""
-        total = self.cache_hits + self.cache_misses
-        return self.cache_hits / total if total > 0 else 0
-
-    def print_stats(self, stream: TextIO = sys.stdout):
-        """Print formatted cache statistics to specified stream
-        Args:
-            stream (TextIO): Output stream (defaults to sys.stdout)
-
-        """
-        stats = f"""
-----------------------------------
-          Cache Statistics
-----------------------------------
-  Number of cache reads:  {self.cache_reads:<10}
-  Number of cache writes: {self.cache_writes:<10}
-  Number of cache hits:   {self.cache_hits:<10}
-  Number of cache misses: {self.cache_misses:<10}
-  Cache hit ratio:        {self.cache_hit_ratio:<9.5%}
-----------------------------------
-"""
-        print(stats, file=stream, flush=True)  # flush avoids buffered output
-
-
-class CacheLogger(Statistics):
-    """
-    Hierarchical logging system with statistics tracking and stream separation.
+    Hierarchical logging system with stream separation.
     Silent logs go to stdout, while Normal and Debug logs go to stderr.
 
     Usage:
@@ -84,7 +24,6 @@ class CacheLogger(Statistics):
         # Custom streams
         > with open('stats.log', 'w') as stdout, open('debug.log', 'w') as stderr:
         >   logger = CacheLogger(level=LogLevel.DEBUG, stdout=stdout, stderr=stderr)
-
     """
 
     def __init__(
@@ -93,7 +32,6 @@ class CacheLogger(Statistics):
         stdout: TextIO = sys.stdout,
         stderr: TextIO = sys.stderr,
     ):
-        super().__init__()
         self.level = level
         self.stdout = stdout
         self.stderr = stderr
@@ -111,20 +49,7 @@ class CacheLogger(Statistics):
         """
         if self.level >= level:
             stream = self.stdout if level == LogLevel.SILENT else self.stderr
-            print(f"{level.name} : {message}", file=stream, flush=True)
-
-    # Statistics recording methods
-    def record_read(self):
-        self.cache_reads += 1
-
-    def record_write(self):
-        self.cache_writes += 1
-
-    def record_hit(self):
-        self.cache_hits += 1
-
-    def record_miss(self):
-        self.cache_misses += 1
+            print(f"{message}", file=stream, flush=True)
 
 
 def log_operation(logger: CacheLogger):
@@ -170,7 +95,6 @@ def log_operation(logger: CacheLogger):
 
     Args:
         logger (CacheLogger): Logger instance that determines logging behavior
-                            and maintains statistics
 
     Returns:
         Callable: A decorator function that will wrap the target function with
@@ -215,7 +139,7 @@ def log_operation(logger: CacheLogger):
                 case "GetSnoopResult":
                     # Address arg[1]
                     logger.log(
-                        LogLevel.NORMAL,
+                        LogLevel.DEBUG,
                         f"GetSnoopResult: Address {args[1]:x}, "
                         f"Snoop Result: {SnoopResult(result).name}",
                     )
@@ -236,7 +160,7 @@ def log_operation(logger: CacheLogger):
 
                 case _:  # Default case
                     logger.log(
-                        LogLevel.NORMAL,
+                        LogLevel.DEBUG,
                         f"[ERROR] Logger could not find matching case for {op_name}, "
                         f"you should not be seeing this",
                     )
